+++ Jumpscale/tools/startupcmd/StartupCMD.py
@@ -9,7 +9,7 @@ class StartupCMD(j.application.JSBaseDataObjClass):
         @url = jumpscale.startupcmd.1
         name* = ""
         cmd_start = ""
-        interpreter = "bash,jumpscale" (E) 
+        interpreter = "bash,jumpscale" (E)
         cmd_stop = ""
         debug = False (b)
         path = ""
@@ -28,7 +28,8 @@ class StartupCMD(j.application.JSBaseDataObjClass):
     @property
     def _pane(self):
         if self._pane_ is None:
-            self._pane_ = j.tools.tmux.pane_get(window=self.name, pane="main", reset=False)
+            self._pane_ = j.tools.tmux.pane_get(
+                window=self.name, pane="main", reset=False)
         return self._pane_
 
     @property
@@ -62,11 +63,12 @@ class StartupCMD(j.application.JSBaseDataObjClass):
     def stop(self):
         self._log_warning("stop:\n%s" % self.name)
         if self.cmd_stop:
-            cmd = j.tools.jinja2.template_render(text=self.cmd_stop, args=self.data._ddict)
+            cmd = j.tools.jinja2.template_render(
+                text=self.cmd_stop, args=self.data._ddict)
             self._log_warning("stopcmd:%s" % cmd)
             rc, out, err = j.sal.process.execute(cmd, die=False)
             time.sleep(0.2)
-        if self.pid > 0:
+        if self.pid and self.pid > 0:
             self._log_info("found process to stop:%s" % self.pid)
             j.sal.process.kill(self.pid)
             time.sleep(0.2)
@@ -114,12 +116,13 @@ class StartupCMD(j.application.JSBaseDataObjClass):
             time.sleep(1)  # need this one or it doesn't check if it failed
         self._log_debug("wait to run:%s (timeout:%s)" % (self.name, timeout))
         while j.data.time.epoch < end:
+            time.sleep(1)
             if self.ports == []:
                 if self.process:
                     if self.process.status().casefold() in ['running', 'sleeping', 'idle']:
                         self._log_info("IS RUNNING %s" % self.name)
                         return True
-                elif self.daemon==False:
+                elif self.daemon == False:
                     return True
             else:
                 nr = 0
@@ -141,13 +144,14 @@ class StartupCMD(j.application.JSBaseDataObjClass):
             self._pane.kill()
         else:
             if self.running:
-                self._log_info("no need to start was already started:%s" % self.name)
+                self._log_info(
+                    "no need to start was already started:%s" % self.name)
                 return
 
         self._pid = None
 
-        if self.interpreter=="bash":
-            C = """         
+        if self.interpreter == "bash":
+            C = """
             reset
             tmux clear
             clear
@@ -160,31 +164,32 @@ class StartupCMD(j.application.JSBaseDataObjClass):
             {% if cmdpath != None %}
             cd {{cmdpath}}
             {% endif %}
-            {{cmd}} 
-    
+            {{cmd}}
+
             """
         else:
             C = """
             from Jumpscale import j
             {% if cmdpath != None %}
             #cd {{cmdpath}}
-            {% endif %}            
-            {{cmd}}          
+            {% endif %}
+            {{cmd}}
             """
-
 
         C2 = j.core.text.strip(C)
         C3 = j.tools.jinja2.template_render(text=C2, args=self.env, cmdpath=self.path,
                                             cmd=self.cmd_start, name=self.name)
-        C3 = C3.replace("\"", "'").replace("''", "'")
-        # for key,val in self.env.items():
+
+        # NEED TO BE CAREFUL< THINGS WILL FAIL IF WE ENABLE AGAIN
+        # if self.interpreter == "bash":
+        #     # C3 = C3.replace("\"", "'").replace("''", "'")
+        #     C3 = C3.replace("\"", "'")
 
         self._log_debug("\n%s" % C3)
 
-
-        if self.interpreter=="bash":
+        if self.interpreter == "bash":
             tpath = self._cmd_path+".sh"
-        elif self.interpreter=="jumpscale":
+        elif self.interpreter == "jumpscale":
             tpath = self._cmd_path+".py"
         else:
             raise RuntimeError("only jumpscale or bash supported")
@@ -195,7 +200,7 @@ class StartupCMD(j.application.JSBaseDataObjClass):
         if "__" in self._pane.name:
             self._pane.kill()
 
-        if self.interpreter=="bash":
+        if self.interpreter == "bash":
             self._pane.execute("source %s" % tpath)
         else:
             if self.debug:
@@ -205,4 +210,4 @@ class StartupCMD(j.application.JSBaseDataObjClass):
 
         if checkrunning:
             running = self.wait_running()
-            assert self.running
+            assert running
